LayoutTests/js/dom/modules/module-incorrect-relative-specifier-expected.txt

CONSOLE MESSAGE: TypeError: Module name, 'incorrect' does not resolve to a valid URL.
CONSOLE MESSAGE: TypeError: Module name, '$hello' does not resolve to a valid URL.
CONSOLE MESSAGE: TypeError: Module name, '.../test' does not resolve to a valid URL.
Test script.onerror will be fired when the incorrect relative specifier is specified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
0
1
2
PASS errors.some(e => e.match(/incorrect/)) is true
PASS errors.some(e => e.match(/hello/)) is true
PASS errors.some(e => e.match(/test/)) is true
PASS Got expected error: 'TypeError'
PASS successfullyParsed is true

TEST COMPLETE
